// src/services/SwapiService/types.ts
export type PlanetResponse = {
  url: string
  name: string
  population: string
  rotation_period: string
  diameter: string
}

export type StarshipResponse = {
  url: string
  name: string
  model: string
  manufacturer: string
  cost_in_credits: string
  length: string
  crew: string
  passengers: string
  cargo_capacity: string
}

export type PersonResponse = {
  url: string
  name: string
  gender: string
  birth_year: string
  eye_color: string
}